+++ next/listtransactionsgrouped/doc.go
@@ -1,0 +1,2 @@
+// Package listtransactionsgrouped shows all transactions grouped by seller
+package listtransactionsgrouped
