--- next/reportgrouped/doc.go
@@ -1,2 +1,0 @@
-// Package reportgrouped shows all transactions grouped by seller
-package reportgrouped
